refactor: posts
