+++ src/app/providers/router/router.tsx
@@ -1,8 +1,8 @@
 import { createBrowserRouter } from "react-router-dom";
 import { Navigate } from "react-router-dom";
 import { AllPostsPage } from "@/pages/AllPostsPage";
-import { PostItemPage } from "@/pages/PostItemPage";
-import { UserPage } from "@/pages/UserPage";
+import { PostItemPage } from "@/pages/PostItemPage/PostItemPage";
+import { UserPage } from "@/pages/UserPage/UserPage";
 import { UserAlbumsPage } from "@/pages/UserAlbumsPage";
 import { UserTodosPage } from "@/pages/UserTodosPage";
 import { UserPostsPage } from "@/pages/UserPostsPage";
@@ -12,7 +12,7 @@ import { NotFoundPage } from "@/pages/NotFoundPage";
 export const router = createBrowserRouter([
   {
     path: "/",
-    element: <Navigate to="/posts" replace />
+    element: <Navigate to="/posts" replace />,
   },
   {
     path: "posts",
