+++ src/pages/PostItemPage/PostItemPage.module.css
@@ -1,0 +1,3 @@
+.goBackBtn {
+  margin-bottom: 20px;
+}
