+++ src/shared/lib/hoc/withLoading.tsx
@@ -2,18 +2,13 @@ import type { ComponentType } from "react";
 import { CircularProgress } from "@mui/material";
 import styles from "./withLoading.module.css";
 
-type withLoadingProps = {
-  loading: boolean;
-  props: any;
-};
-
-export function withLoading(Component: ComponentType) {
-  return ({ loading, props }: withLoadingProps) =>
+export function withLoading<P extends object>(Component: ComponentType<P>) {
+  return ({ loading, ...props }: { loading: boolean } & P) =>
     loading ? (
       <div className={styles.loadingContainer}>
         <CircularProgress />
       </div>
     ) : (
-      <Component {...props} />
+      <Component {...(props as P)} />
     );
 }
